requirements.txt
opencv-python
numpy
Pillow
watchdog

